update list function with backfill and ensure data from graph and kvdb
